% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doit.R
\name{doit_approx}
\alias{doit_approx}
\title{DoIt approximation of the target density}
\usage{
doit_approx(doit, theta_eval)
}
\arguments{
\item{doit}{An object of class `doit`, see function `doit_fit`.}

\item{theta_eval}{A data frame of evaluation points.}
}
\value{
A data frame of the provided evaluation points and the corresponding
mean and variance of the DoIt approximation.
}
\description{
Evaluate mean and variance of the DoIt approximation of the target density
at a number of evaluation points.
}
