% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doit.R
\name{doit_estimate_w}
\alias{doit_estimate_w}
\title{Estimate kernel widths for DoIt}
\usage{
doit_estimate_w(design, w_0 = NULL, optim_control = NULL)
}
\arguments{
\item{design}{A data frame of design points and corresponding function
evaluations. Must contain a column named `f` with function values. The other
columns are treated as design points.}

\item{w_0}{Starting point for the optimisation routine. Either NULL
(default) in wwhich case an educated guess is made, or otherwise a vector of
the same dimension as the parameter space.}

\item{optim_control}{A list passed to `optimize` (for 1d problems) or
`optim` (if the parameter has dimension 2 or more)}
}
\value{
A vector of optimal kernel widths.
}
\description{
Estimate the optimal kernel bandwith by minimising the weighted
mean squared cross validation error.
}
