% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doit.R
\name{doit_expectation}
\alias{doit_expectation}
\title{DoIt approximation of posterior expectation}
\usage{
doit_expectation(doit)
}
\arguments{
\item{doit}{An object of class `doit`, see function `doit_fit`.}
}
\value{
Vector of posterior expectations
}
\description{
DoIt approximation of posterior expectation
}
