% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doit.R
\name{doit_fit}
\alias{doit_fit}
\title{Fit a DoIt object}
\usage{
doit_fit(design, sigma2 = NULL)
}
\arguments{
\item{design}{A data frame of design points and corresponding function
evaluations. Must contain a column named `f` with function values. The other
columns are treated as design points.}

\item{sigma2}{vector of variances used for the Gaussian kernels. If `NULL`
(the default), sigma2 is calculated by `doit_estimate_sigma`.}
}
\value{
An object of class `doit` used in further calculations.
}
\description{
Fit the parameters of the DoIt approximation.
}
