% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doit.R
\name{doit_propose_new}
\alias{doit_propose_new}
\title{Propose a new design point for the DoIt approximation}
\usage{
doit_propose_new(doit, near = NULL)
}
\arguments{
\item{doit}{An object of class `doit`, see function `doit_fit`.}

\item{theta_0}{Data frame or vector. Initial value used by the optimisation
routine. Ignored for 1d problems.}
}
\value{
A parameter value.
}
\description{
Find the point with largest estimation variance by numerical optimisation,
using the current design point with largest leave-one-out prediction error
as the starting point.
}
\examples{
design = data.frame(x=rnorm(10), y=rnorm(10))
design$f = with(design, exp(-0.5*(x+y)^2))
fit = doit_fit(design)
theta_new = doit_propose_new(fit)

}
