% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doit.R
\name{doit_update}
\alias{doit_update}
\title{Update a DoIt approximation}
\usage{
doit_update(doit, design_new)
}
\arguments{
\item{doit}{An object of class `doit`, see function `doit_fit`.}

\item{design_new}{The new design point.}
}
\value{
The updated `doit` object.
}
\description{
Update a `doit` object by adding a new design point, but without changing
the kernel width.
}
\examples{
design = data.frame(x=rnorm(10), y=rnorm(10))
design$f = with(design, exp(-0.5*(x+y)^2))
fit = doit_fit(design)
theta_new = doit_propose_new(fit)
theta_new$f = with(theta_new, exp(-0.5*(x+y)^2))
fit = doit_update(fit, theta_new)

}
