% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doit.R
\name{doit_variance}
\alias{doit_variance}
\title{DoIt approximation of the variance}
\usage{
doit_variance(doit)
}
\arguments{
\item{doit}{An object of class `doit`, see function `doit_fit`.}
}
\value{
(Co-)variance matrix of the target density.
}
\description{
Approximate the covariance matrix of the vector `theta`.
}
\examples{
design = data.frame(x=rnorm(10), y=rnorm(10))
design$f = with(design, exp(-0.5*(x+y)^2))
fit = doit_fit(design)
vari = doit_variance(fit)

}
