% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doit.R
\name{doit_approx}
\alias{doit_approx}
\title{DoIt approximation of the target density}
\usage{
doit_approx(doit, theta_eval)
}
\arguments{
\item{doit}{An object of class `doit`, see function `doit_fit`.}

\item{theta_eval}{A data frame of evaluation points.}
}
\value{
A data frame of the provided evaluation points and the corresponding
mean and variance of the DoIt approximation.
}
\description{
Evaluate mean and variance of the DoIt approximation of the target density
at a number of evaluation points.
}
\examples{
design = data.frame(x=rnorm(10), y=rnorm(10))
design$f = with(design, exp(-0.5*(x+y)^2))
fit = doit_fit(design)
theta_eval = expand.grid(x=seq(-2,2,.1), y=seq(-2,2,.1))
approx = doit_approx(fit, theta_eval)

}
