% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doit.R
\name{doit_expectation}
\alias{doit_expectation}
\title{DoIt approximation of the expectation}
\usage{
doit_expectation(doit)
}
\arguments{
\item{doit}{An object of class `doit`, see function `doit_fit`.}
}
\value{
Vector of expected values of the target density.
}
\description{
Approximate the expectation of the vector `theta`.
}
