% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doit.R
\name{doit_expectation}
\alias{doit_expectation}
\title{DoIt approximation of the expectation}
\usage{
doit_expectation(doit)
}
\arguments{
\item{doit}{An object of class `doit`, see function `doit_fit`.}
}
\value{
Vector of expected values of the target density.
}
\description{
Approximate the expectation of the vector `theta`.
}
\examples{
design = data.frame(x=rnorm(10), y=rnorm(10))
design$f = with(design, exp(-0.5*(x+y)^2))
fit = doit_fit(design)
expec = doit_expectation(fit)

}
