% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doit.R
\name{doit_integral}
\alias{doit_integral}
\title{DoIt approximation of the normalisation constant}
\usage{
doit_integral(doit)
}
\arguments{
\item{doit}{An object of class `doit`, see function `doit_fit`.}
}
\value{
The integral of the approximated function over all `theta`.
}
\description{
Calculate the integral of the approximated function.
}
