% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doit.R
\name{doit_marginal_A}
\alias{doit_marginal_A}
\title{DoIt approximation for a linear transformations}
\usage{
doit_marginal_A(doit, A = NULL, theta_eval = NULL)
}
\arguments{
\item{doit}{An object of class `doit`, see function `doit_fit`.}

\item{A}{The transformation matrix, with number of columns equal to the
number of parameters. Default is the identity matrix. The rownames of `A`
are used to name the elements of the transformation. If unspecified, row
names are set to `tau1`, `tau2`, ...}

\item{theta_eval}{Parameter values at which to approximate the density of
the transformation. If `NULL` (the default) the original design points are
used.}
}
\value{
A data frame of the transformed evaluation points and the
corresponding DoIt approximation of the marginal density.
}
\description{
Approximate the marginal density of a linear transformation `tau = A %*%
theta`, such as `tau = theta[1]`, `tau = theta[1:3]` or `tau =
theta[1]+theta[2]`. If `tau` has dimension greater than 1, the joint
marginal density is approximated.
}
