% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doit.R
\name{doit_marginals}
\alias{doit_marginals}
\title{DoIt approximation of all marginal posterior distributions.}
\usage{
doit_marginals(doit, theta_eval = NULL)
}
\arguments{
\item{doit}{An object of class `doit`, see function `doit_fit`.}

\item{theta_eval}{Evaluation points at which to approximate the marginal
distribution. If `NULL` (the default) the original design points are used.}
}
\value{
A data frame of the provided evaluation points and the corresponding
DoIt approximation of the marginal posterior.
}
\description{
DoIt approximation of all marginal posterior distributions.
}
