% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doit.R
\name{doit_propose_new}
\alias{doit_propose_new}
\title{Propose a new design point for the DoIt approximation}
\usage{
doit_propose_new(doit)
}
\arguments{
\item{doit}{An object of class `doit`, see function `doit_fit`.}
}
\value{
A parameter value.
}
\description{
Find the point with largest estimation variance by numerical optimisation,
using the current design point with largest leave-one-out prediction error
as the starting point.
}
