% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doit.R
\name{doit_update}
\alias{doit_update}
\title{Update a DoIt approximation}
\usage{
doit_update(doit, design_new)
}
\arguments{
\item{doit}{An object of class `doit`, see function `doit_fit`.}

\item{design_new}{The new design point.}
}
\value{
The updated `doit` object.
}
\description{
Update a `doit` object by adding a new design point, but without changing
the kernel width.
}
