% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doit.R
\name{doit_variance}
\alias{doit_variance}
\title{DoIt approximation of posterior variance}
\usage{
doit_variance(doit)
}
\arguments{
\item{doit}{An object of class `doit`, see function `doit_fit`.}
}
\value{
Posterior (co-)variance matrix.
}
\description{
DoIt approximation of posterior variance
}
